# backend/test/test_from_swagger/swagger.json
{
    "components": {
        "schemas": {
...
            "InvestmentTransaction": {
                "properties": {
                    "activity_type": {
                        "enum": [
                            "buy",
                            "sell",
                            "deposit",
                            "withdrawal"
                        ],
                        "type": "string"
                    },
                    "asset_id": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string"
                    },
                    "fee": {
                        "type": "number"
                    },
                    "from_account_id": {
                        "type": "integer"
                    },
                    "id": {
                        "readOnly": true,
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "tax": {
                        "type": "number"
                    },
                    "to_account_id": {
                        "type": "integer"
                    },
                    "total_paid": {
                        "readOnly": true,
                        "type": "number"
                    },
                    "unit_price": {
                        "type": "number"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                },
                "required": [
                    "activity_type",
                    "asset_id",
                    "date",
                    "fee",
                    "from_account_id",
                    "quantity",
                    "tax",
                    "to_account_id",
                    "total_paid",
                    "unit_price",
                    "user_id"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "bearerFormat": "JWT",
                "description": "Enter your JWT token in the format: Bearer <token>",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "contact": {
            "email": "support@wealthmanager.com"
        },
        "description": "API documentation for the WealthManager application",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "title": "WealthManager API",
        "version": "1.0.0"
    },
    "openapi": "3.0.2",
    "paths": {
...
        "/investments/": {
            "get": {
                "parameters": [
                    {
                        "description": "Page number",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Items per page",
                        "in": "query",
                        "name": "per_page",
                        "schema": {
                            "default": 10,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Field to sort by",
                        "in": "query",
                        "name": "sort_by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Sort order (asc or desc)",
                        "in": "query",
                        "name": "sort_order",
                        "schema": {
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated list of fields to return",
                        "in": "query",
                        "name": "fields",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Search query",
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "items": {
                                                "$ref": "#/components/schemas/InvestmentTransaction"
                                            },
                                            "type": "array"
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "pages": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "List of investments"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get all investments",
                "tags": [
                    "Investment"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvestmentTransaction"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvestmentTransaction"
                                }
                            }
                        },
                        "description": "Investment created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create a new investment",
                "tags": [
                    "Investment"
                ]
            }
        },
        "/investments/assets/{symbol}/transactions": {
            "get": {
                "parameters": [
                    {
                        "description": "Asset symbol (e.g., AAPL, GOOGL)",
                        "in": "path",
                        "name": "symbol",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "date": {
                                                "format": "date",
                                                "type": "string"
                                            },
                                            "fees": {
                                                "type": "number"
                                            },
                                            "id": {
                                                "type": "integer"
                                            },
                                            "price": {
                                                "type": "number"
                                            },
                                            "shares": {
                                                "type": "number"
                                            },
                                            "total": {
                                                "type": "number"
                                            },
                                            "type": {
                                                "enum": [
                                                    "buy",
                                                    "sell"
                                                ],
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Asset transactions"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get transactions for a specific investment asset",
                "tags": [
                    "Investment"
                ]
            }
        },
        "/investments/portfolio/performance": {
            "get": {
                "parameters": [
                    {
                        "description": "Time period for performance data",
                        "in": "query",
                        "name": "period",
                        "required": false,
                        "schema": {
                            "default": "1Y",
                            "enum": [
                                "1D",
                                "1W",
                                "1M",
                                "3M",
                                "6M",
                                "1Y",
                                "3Y",
                                "5Y",
                                "ALL"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data_points": {
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "format": "date",
                                                        "type": "string"
                                                    },
                                                    "value": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "type": "array"
                                        },
                                        "end_value": {
                                            "type": "number"
                                        },
                                        "period": {
                                            "type": "string"
                                        },
                                        "start_value": {
                                            "type": "number"
                                        },
                                        "total_return": {
                                            "type": "number"
                                        },
                                        "total_return_percentage": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Portfolio performance data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get investment portfolio performance over time",
                "tags": [
                    "Investment"
                ]
            }
        },
        "/investments/portfolio/summary": {
            "get": {
                "parameters": [
                    {
                        "description": "Optional account ID to filter portfolio summary",
                        "in": "query",
                        "name": "account_id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "assets": {
                                            "items": {
                                                "properties": {
                                                    "cost_basis": {
                                                        "type": "number"
                                                    },
                                                    "current_price": {
                                                        "type": "number"
                                                    },
                                                    "current_value": {
                                                        "type": "number"
                                                    },
                                                    "gain_loss": {
                                                        "type": "number"
                                                    },
                                                    "gain_loss_percentage": {
                                                        "type": "number"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "shares": {
                                                        "type": "number"
                                                    },
                                                    "symbol": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "type": "array"
                                        },
                                        "total_cost": {
                                            "type": "number"
                                        },
                                        "total_gain_loss": {
                                            "type": "number"
                                        },
                                        "total_gain_loss_percentage": {
                                            "type": "number"
                                        },
                                        "total_value": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Portfolio summary data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get investment portfolio summary",
                "tags": [
                    "Investment"
                ]
            }
        },
        "/investments/{id}": {
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Delete a investment",
                "tags": [
                    "Investment"
                ]
            },
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvestmentTransaction"
                                }
                            }
                        },
                        "description": "Return investment"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get a investment by ID",
                "tags": [
                    "Investment"
                ]
            },
            "put": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvestmentTransaction"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvestmentTransaction"
                                }
                            }
                        },
                        "description": "Investment updated successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update a investment",
                "tags": [
                    "Investment"
                ]
            }
        },
...
    },
    "servers": [
        {
            "description": "Development server",
            "url": "https://wealthmanager.myfunnycluster.dynamic-dns.net"
        }
    ]
}